requirements.txt
python-dotenv==0.19.2
discord-py-slash-command==3.0.3
discord.py==1.7.3
mcstatus==6.5.0
